==> notebooks/01_preprocess.ipynb <==
**Imports**

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
import re
import string
from nltk.corpus import stopwords
from nltk.stem import PorterStemmer

**Text Cleaning Util**

stop_words = set(stopwords.words('english'))
stemmer = PorterStemmer()

def clean_text(text):
    text = text.lower()
    text = re.sub(r'\d+', '', text)
    text = re.sub(r'http\S+|www\S+|https\S+', '', text, flags=re.MULTILINE)
    # text = text.translate(str.maketrans('', '', string.punctuation))
    text = re.sub(r'[^a-zA-Z\s]', '', text)
    text = ' '.join([word for word in text.split() if word not in stop_words])
    text = ' '.join([stemmer.stem(word) for word in text.split()])
    text = re.sub(r'\s+', ' ', text).strip()
    return text


## SPAM

df = pd.read_csv("..\data\orig\main\spam\data.csv", encoding='ISO-8859-1')
df.head()

df = df[["v1", "v2"]]
df["v1"] = df["v1"].apply(lambda x: 1 if x=="spam" else 0)
df["v2"] = df["v2"].apply(clean_text)
df.rename(columns={"v1":"y","v2":"text"}, inplace=True)

df = df.sample(frac=1).reset_index(drop=True)
df_train, df_test = train_test_split(df, test_size=0.2, random_state=65)

df_train.to_csv("../data/orig/processed/train/spam-data.csv", index=False)
df_test.to_csv("../data/orig/processed/test/spam-data.csv", index=False)

#check
df_train = pd.read_csv("../data/orig/processed/train/spam-data.csv")
df_train['y'].value_counts()

## SENTIMENT

df = pd.read_csv("..\data\orig\main\sentiment\data.csv", encoding='latin-1', header=None)
df.head()

df = df[[0,5]]
df.rename(columns={0:'y',5:'text'}, inplace=True)
df['y'] = df['y'].apply(lambda x: 1 if x==4 else 0)
df['text'] = df['text'].apply(clean_text) #this takes 5mins+
df = df.sample(frac=1).reset_index(drop=True)

df_train, df_test = train_test_split(df, random_state=46, test_size=0.3)

df_train.to_csv("../data/orig/processed/train/sentiment-data.csv", index=False)
df_test.to_csv("../data/orig/processed/test/sentiment-data.csv", index=False)

## NEWS

df_1 = pd.read_csv("..\\data\\orig\\main\\news\\BuzzFeed_fake_news_content.csv")
df_2 = pd.read_csv("..\\data\\orig\\main\\news\\BuzzFeed_real_news_content.csv")

print("No. of Rows (Fake): ", df_1.shape[0])
print("No. of Rows (Real): ", df_2.shape[0])

df_1['text'] = df_1['title'] + ' ' + df_1['text']
df_2['text'] = df_2['title'] + ' ' + df_2['text']

df_1 = df_1[['text']]
df_2 = df_2[['text']]

df_1['text'] = df_1['text'].apply(clean_text)
df_2['text'] = df_2['text'].apply(clean_text)

df_1['y'] = 1
df_2['y'] = 0

df1_train, df1_test = train_test_split(df_1, random_state=42)
df2_train, df2_test = train_test_split(df_2, random_state=36)

df_train = pd.concat([df1_train, df2_train], ignore_index=True)
df_train = df_train.sample(frac=1).reset_index(drop=True)
df_train.to_csv("../data/orig/processed/train/news-data.csv", index=False)

df_test = pd.concat([df1_test, df2_test], ignore_index=True)
df_test = df_test.sample(frac=1).reset_index(drop=True)
df_test.to_csv("../data/orig/processed/test/news-data.csv", index=False)

**Minis**


df_spam = pd.read_csv("..\\data\\orig\\processed\\train\\spam-data.csv")
df_spam['y'].value_counts()

df_spam = pd.concat([df_spam[df_spam['y']==1], df_spam[df_spam['y']==0].sample(n=602)], ignore_index=True)

df_spam.to_csv("..\\data\\orig\\processed\\train\\spam-data-mini.csv")
df_spam.shape

df_sentiment = pd.read_csv("..\\data\\orig\\processed\\train\\sentiment-data.csv")
df_sentiment.shape

df_sentiment['y'].value_counts()

n = 2500
df_sentiment = pd.concat([df_sentiment[df_sentiment['y']==1].sample(n=n), df_sentiment[df_sentiment['y']==0].sample(n=n)], ignore_index=True)

df_sentiment.to_csv("..\\data\\orig\\processed\\train\\sentiment-data-mini.csv")
df_sentiment.shape
