==> notebooks/02_4_fsl_prompt_gen.ipynb <==
import pandas as pd
import numpy as np


GPT2 = "gpt2"
FLAN = "flan"

from transformers import AutoModelForSeq2SeqLM, AutoTokenizer

model_name = "google/flan-t5-small"
tokenizer = AutoTokenizer.from_pretrained(model_name)
model = AutoModelForSeq2SeqLM.from_pretrained(model_name) 

def generate_text_from_prompt(prompt):
    inputs = tokenizer(prompt, return_tensors="pt", max_length=49)
    
    outputs = model.generate(**inputs, max_length=50, num_return_sequences=1, do_sample=True, top_k=0)
    generated_text = tokenizer.decode(outputs[0], skip_special_tokens=True)
    return generated_text

from transformers import GPT2LMHeadModel, GPT2Tokenizer

model_name = "gpt2"
tokenizer = GPT2Tokenizer.from_pretrained(model_name)
model = GPT2LMHeadModel.from_pretrained(model_name, pad_token_id=tokenizer.eos_token_id)

def generate_text_from_prompt(prompt):
    inputs = tokenizer(prompt, return_tensors="pt", max_length=49)
    
    outputs = model.generate(**inputs, max_length=50, num_return_sequences=1, do_sample=True, top_k=0)
    generated_text = tokenizer.decode(outputs[0], skip_special_tokens=True)
    return generated_text

def get_fsl_data(task):
    return pd.read_csv("../data/syn/mid/fsl-cat-"+task+".csv")

tasks = {
    'news':{
        'loop':10
    },
    'spam':{
        'loop':100
    },
    'sentiment':{
        'loop':250
    }
}

LLM = FLAN

# df = get_fsl_data(task)
# prompt = df[df['fsl_category']==0].sample(n=1)
# prompt['text'].values[0]

for task in tasks:
    df = get_fsl_data(task)
    df0_n = pd.DataFrame(columns=['text'])
    df1_n = pd.DataFrame(columns=['text'])
    #cat 0
    df0 = df[df['y']==0]
    print(f"Start: {task}")
    print(f"df0 rows: {df0.shape[0]}, nun: {df0['fsl_category'].nunique()}")
    for value, count in df0['fsl_category'].value_counts().items(): 
        print(f"\nValue: {value}, Count: {count}")
        prompt = df0[df0['fsl_category']==value].sample(n=1)
        prompt = prompt['text'].values[0]
        print(f"Prompt: {prompt[:20]}")
        for _ in range(count):
            text = generate_text_from_prompt(prompt)
            df0_n.loc[len(df0_n)] = text
            if _%tasks[task]['loop']==0:
                print(_, end=" ")
        print(f"\nFinished Cat: ", value)
    df0_n['y']=0
    print("Finished df0")

    df1 = df[df['y']==0]
    print(f"Start: {task}")
    print(f"df1 rows: {df1.shape[0]}, nun: {df1['fsl_category'].nunique()}")
    for value, count in df0['fsl_category'].value_counts().items(): 
        print(f"\nValue: {value}, Count: {count}")
        prompt = df1[df1['fsl_category']==value].sample(n=1)
        prompt = prompt['text'].values[0]
        print(f"Prompt: {prompt[:20]}")
        for _ in range(count):
            text = generate_text_from_prompt(prompt)
            df1_n.loc[len(df1_n)] = text
            if _%tasks[task]['loop']==0:
                print(_, end=" ")
        print(f"\nFinished Cat: ", value)
    df1_n['y']=1
    print("Finished df1")

    df = pd.concat([df1_n, df0_n], ignore_index=True).sample(frac=1)
    df.to_csv('../data/syn/'+LLM+'/fs/auto-'+task+'-fs.csv', index=False)
