==> notebooks/02_5_flan_pp_gen.ipynb <==
**Imports**

import pandas as pd
from transformers import AutoModelForSeq2SeqLM, AutoTokenizer
import torch
from sklearn.model_selection import train_test_split

**Model: FLAN**

model_name = "google/flan-t5-small"
tokenizer = AutoTokenizer.from_pretrained(model_name)
model = AutoModelForSeq2SeqLM.from_pretrained(model_name) 


**Utils**

def count_rows(task):
    file = "../data/orig/processed/train/" + task + "-data-mini.csv"
    df = pd.read_csv(file)
    return int(df.shape[0]/2)

def load_data(task):
    if(task == "news"):
        df_1 = pd.read_csv("..\\data\\orig\\main\\news\\BuzzFeed_fake_news_content.csv")
        df_2 = pd.read_csv("..\\data\\orig\\main\\news\\BuzzFeed_real_news_content.csv")

        df_1['text'] = df_1['title'] + ' ' + df_1['text']
        df_2['text'] = df_2['title'] + ' ' + df_2['text']

        df_1 = df_1[['text']]
        df_2 = df_2[['text']]

        df_1['y'] = 1
        df_2['y'] = 0

        df1_train, df1_test = train_test_split(df_1, random_state=42)
        df2_train, df2_test = train_test_split(df_2, random_state=36)

        df1 = pd.concat([df1_train, df2_train], ignore_index=True)
        df1 = df1.sample(frac=1).reset_index(drop=True)
        return df1
    
    if(task == "spam"):
        df = pd.read_csv("../data/orig/main/spam/data.csv", encoding='ISO-8859-1')
        df = df[["v1", "v2"]]
        df["v1"] = df["v1"].apply(lambda x: 1 if x=="spam" else 0)
        df.rename(columns={"v1":"y","v2":"text"}, inplace=True)
        df = df.sample(frac=1).reset_index(drop=True)
        df1, df2 = train_test_split(df, test_size=0.2, random_state=65)
        df1.dropna(inplace=True)
        df1 = pd.concat([df1[df1['y']==1], df1[df1['y']==0].sample(n=602)], ignore_index=True)
        return df1
    df = pd.read_csv("../data/orig/main/sentiment/data.csv", encoding='latin-1', header=None)
    df = df[[0,5]]
    df.rename(columns={0:'y',5:'text'}, inplace=True)
    df['y'] = df['y'].apply(lambda x: 1 if x==4 else 0)
    df = df.sample(frac=1).reset_index(drop=True)
    df1, df2  = train_test_split(df, random_state=46, test_size=0.3)
    df1.dropna(inplace=True)
    n = 2500
    df1 = pd.concat([df1[df1['y']==1].sample(n=n), df1[df1['y']==0].sample(n=n)], ignore_index=True)
    return df1

zsl_tasks = [
    {
        'name':'news',
        'rows': count_rows('news'),
        'print_count':10
    },
    {
        'name':'spam',
        'rows': count_rows('spam'),
        'print_count':100
    },    
    {
        'name':'sentiment',
        'rows': count_rows('sentiment'),
        'print_count':250
    }
]

def generate_text_from_prompt(prompt):
    inputs = tokenizer(prompt, return_tensors="pt", max_length=49)
    
    outputs = model.generate(**inputs, max_length=50, num_return_sequences=1, do_sample=True, top_k=0)
    generated_text = tokenizer.decode(outputs[0], skip_special_tokens=True)
    return generated_text

# generate_text_from_prompt(prompt="is upset that he can't update his Facebook by texting it... and might cry as a result  School today also. Blah!")


**ZSP - Data Generation**

import re
import string
from nltk.corpus import stopwords
from nltk.stem import PorterStemmer

stop_words = set(stopwords.words('english'))
stemmer = PorterStemmer()

def clean_text(text):
    text = text.lower()
    text = re.sub(r'\d+', '', text)
    text = re.sub(r'http\S+|www\S+|https\S+', '', text, flags=re.MULTILINE)
    # text = text.translate(str.maketrans('', '', string.punctuation))
    text = re.sub(r'[^a-zA-Z\s]', '', text)
    text = ' '.join([word for word in text.split() if word not in stop_words])
    text = ' '.join([stemmer.stem(word) for word in text.split()])
    text = re.sub(r'\s+', ' ', text).strip()
    return text


df_main = pd.DataFrame()
for task in zsl_tasks:
    print("1. Starting task: ", task)    
    df = load_data(task['name'])
    print("2. Loaded Data, Rows: ", df.shape[0])
    print("3. Prog: ")
    for _ in range(df.shape[0]):
        new_text = generate_text_from_prompt(prompt=df.loc[_,'text'])
        df.loc[_,'text'] = clean_text(new_text)
        df_main = df
        if _%task['print_count'] == 0:
            print(_, end=' ')
            df.to_csv("../data/syn/mid/auto-" + task['name'] + "-pp-data.csv", index=False)
    print("\n4. Saving df")
    df = df.sample(frac=1).reset_index(drop=True)
    df.dropna(inplace=True)
    df_main = df
    df.to_csv("../data/syn/flan/pp/auto-" + task['name'] + "-data.csv", index=False)
    print("5. Ending task: ", task['name'])
    print()
